+++ createTempDir.mjs
@@ -1,6 +1,6 @@
 // @ts-check
 
-import fs from "fs";
+import { mkdtemp, realpath } from "fs/promises";
 import { tmpdir } from "os";
 import { sep } from "path";
 
@@ -11,6 +11,6 @@ import { sep } from "path";
 export default async function createTempDir() {
   // On macOS `os.tmpdir()` returns the path to a symlink, see:
   // https://github.com/nodejs/node/issues/11422
-  const osTempDirPath = await fs.promises.realpath(tmpdir());
-  return fs.promises.mkdtemp(`${osTempDirPath}${sep}`);
+  const osTempDirPath = await realpath(tmpdir());
+  return mkdtemp(`${osTempDirPath}${sep}`);
 }