+++ createTempDir.test.mjs
@@ -2,9 +2,9 @@
 
 import { strictEqual } from "assert";
 import { existsSync } from "fs";
+import { rm } from "fs/promises";
 
 import createTempDir from "./createTempDir.mjs";
-import fsPathRemove from "./fsPathRemove.mjs";
 
 /**
  * Adds `createTempDir` tests.
@@ -14,6 +14,6 @@ export default (tests) => {
   tests.add("`createTempDir`.", async () => {
     const tempDirPath = await createTempDir();
     strictEqual(existsSync(tempDirPath), true);
-    await fsPathRemove(tempDirPath);
+    await rm(tempDirPath, { force: true, recursive: true });
   });
 };